% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_df.R
\name{build_df}
\alias{build_df}
\title{Build a data frame of building parameters}
\usage{
build_df(run_bpip = "N", height = as.numeric(NA),
  width_x = as.numeric(NA), length_y = as.numeric(NA),
  bld_rotation = as.numeric(NA), dist_from_source = as.numeric(NA),
  angle_from_source = as.numeric(NA))
}
\arguments{
\item{run_bpip}{Building ID or name.}

\item{height}{Height of building in meters.}

\item{width_x}{Width of building from East to West in meters.}

\item{length_y}{Length of building from North to South in meters.}

\item{bld_rotation}{Building rotation from North in degrees.}

\item{dist_from_source}{Distance from center of source to center of building in meters.}

\item{angle_from_source}{Angle between center of building and center of source relative to North.
Units in degrees. 
Straight North is "0" degrees.
Straight East is "90" degrees.}
}
\description{
Create an input table of building parameters for AERSCREEN.
}
\examples{
build_df(run_bpip = "Y")
}
\keyword{aerscreen}
\keyword{bpip}
\keyword{building}
\keyword{input}

