% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_aerscreen.R
\name{install_aerscreen}
\alias{install_aerscreen}
\title{Download AERSCREEN.exe from EPA}
\usage{
install_aerscreen(folder = getwd(), aermap = TRUE, aermod = TRUE,
  bpip = TRUE, makemet = TRUE)
}
\arguments{
\item{folder}{Download location. Default is the current working directory.}

\item{aermap}{Download AERMAP? A logical value indicating whether to download EPA's terrain preprocessor AERMAP.}

\item{bpip}{Download BPIPPRM? A logical value indicating whether to download EPA's downwash calculator BPIPPRM.}

\item{makemet}{Download MAKEMET? A logical value indicating whether to download EPA's simplified meteorology preprocessor MAKEMET.}
}
\description{
Download AERSCREEN.exe from EPA. Option to install AERMAP, AERMOD, BPIPPRM, and MAKEMET.
}
\examples{
install_aerscreen(folder = "my_facility")
}
\keyword{aermod}
\keyword{aerscreen}
\keyword{bpip}
\keyword{download}
\keyword{install}
\keyword{makemet}

