% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aerscreen_inp.R
\name{read_aerscreen_inp}
\alias{read_aerscreen_inp}
\title{Read AERSCREEN input file}
\usage{
read_aerscreen_inp(file = "aerscreen.inp")
}
\arguments{
\item{file}{A path to a file or literal data as a character string. Default is "aerscreen.inp" in the working directory.}
}
\description{
Read an aerscreen.inp file or text into an AERSCREEN input table.
}
\examples{
read_aerscreen_inp(file = "aermod.inp")
}
\keyword{aerscreen}
\keyword{input}
\keyword{load}
\keyword{read}

