% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aerscreen_out.R
\name{read_aerscreen_out}
\alias{read_aerscreen_out}
\title{Read AERSCREEN output file}
\usage{
read_aerscreen_out(path)
}
\arguments{
\item{path}{A path to a file or data as a character string.}
}
\description{
Read an aerscreen.out file into an AERSCREEN results table.
}
\examples{
\dontrun{
read_aerscreen_out(file = "aerscreen.out")
}
}
\keyword{aerscreen}
\keyword{load}
\keyword{output}
\keyword{read}
\keyword{results}

