% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aerscreen.R
\name{run_aerscreen}
\alias{run_aerscreen}
\title{Run AERSCREEN}
\usage{
run_aerscreen(input = "aerscreen.inp", output = input$TITLEONE,
  exe_folder = "aerscreen/")
}
\arguments{
\item{input}{AERSCREEN input file. Defaults to aerscreen_inp. A file path may be used (ex. "aerscreen.inp").}

\item{output}{File name for model output. Defaults to `TITLEONE` of input file.}

\item{exe_folder}{aerscreen.exe location. Defaults to "aerscreen/".}
}
\description{
Run aerscreen.exe with designated input file.
}
\examples{
run_aerscreen(input = "inputs/aerscreen.inp", 
              output = "output_test1")
}
\keyword{aermod}
\keyword{aerscreen}
\keyword{dispersion}
\keyword{model}

