% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_df.R
\name{surface_df}
\alias{surface_df}
\title{Surface characteristics}
\usage{
surface_df(min_temp_k = 235, max_temp_k = 315, min_wind_speed = 0.5,
  anem_height = 10, surface_profile = 7, climate_profile = 1,
  albedo = as.numeric(NA), bowen = as.numeric(NA),
  z_length = as.numeric(NA))
}
\arguments{
\item{min_temp_k}{Minimum surface temperature in Kelvin. Defaults to \code{235}.}

\item{max_temp_k}{Maximum surface temperature in Kelvin. Defaults to \code{315}.}

\item{min_wind_speed}{Minimum wind speed in units of \code{m/s}. 
Requires a value of \code{0.5} or greater. 
Defaults to \code{0.5}.}

\item{anem_height}{Anemometer height in meters. Defaults to \code{10}.}

\item{surface_profile}{Dominant surface characteristics profile. Accepts integers between [0, 3].
Enter \code{0} if providing site specific albedo, Bowen ratio and surface roughness.
Defaults to \code{7}. Options: 
(0) None. User specified characteristics will be used.
(1) Water
(2) Deciduous forest
(3) Coniferous forest
(4) Swamp
(5) Cultivated land
(6) Grassland
(7) Urban
(8) Desert shrubland}

\item{climate_profile}{Dominant climate moisture profile. Accepts integers between [0, 3]. 
Enter \code{0} if providing site specific albedo, Bowen ratio and surface roughness. 
Defaults to \code{1}. Options: 
(0) None. User specified characteristics will be used. 
(1) Average moisture
(2) Wet conditions
(3) Dry conditions}

\item{albedo}{Albedo. The average proportion of incident light reflected at noontime.
Accepts values between [0, 1].
Ignored by AERSCREEN unless \code{surface_profile} and \code{climate_profile} are set to \code{0}.}

\item{bowen}{Bowen ratio. The average midday ratio of the sensible heat flux (\code{H}) 
to the evaporative heat flux (\code{E}). 
The greater the value of \code{H/E} the dryer the conditions.
Accepts values between [-10 and 10].
Ignored by AERSCREEN unless \code{surface_profile} and \code{climate_profile} are set to \code{0}.}

\item{z_length}{Surface roughness length. 
The height at  which the  mean  horizontal  wind  speed  approaches  zero. 
Accepts values between [0.001 and 2].
Ignored by AERSCREEN unless \code{surface_profile} and \code{climate_profile} are set to \code{0}.}
}
\description{
Create a data frame of AERSCREEN surface characteristics used by MAKEMET.
}
\examples{
surface_df()
}
\keyword{aerscreen}
\keyword{input}
\keyword{meteorology}
\keyword{surface}
\keyword{weather}

