% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_aerscreen.R
\name{new_aerscreen}
\alias{new_aerscreen}
\title{Create AERSCREEN input tables}
\usage{
new_aerscreen(aerscreen = "aerscreen_inp", as_one_df = TRUE,
  add_to_envir = FALSE, control = "control", sources = "sources",
  buildings = "buildings", surface = "surface")
}
\arguments{
\item{aerscreen}{Name for AERSCREEN input data frame added to Global Environment. 
Default is "aerscreen_inp". Ignored if \code{as_one_df} is \code{FALSE}.}

\item{as_one_df}{\code{TRUE} or \code{FALSE}. 
Return all inputs in a single wide data frame. 
If \code{FALSE}, return 4 data frames in a list:
    (1) control 
    (2) sources
    (3) buildings
    (4) surface}

\item{add_to_envir}{\code{TRUE} or \code{FALSE}. 
Exports tables directly to the Global Environment.}

\item{control}{Name for control options data frame added to Global Environment. 
Default is "sources". Ignored if \code{as_one_df} is \code{TRUE}.}

\item{sources}{Name for emission source data frame added to Global Environment. 
Default is "sources". Ignored if \code{as_one_df} is \code{TRUE}.}

\item{buildings}{Name for buildings data frame added to Global Environment. 
Default is "buildings". Ignored if \code{as_one_df} is \code{TRUE}.}

\item{surface}{Name for surface characteristics data frame added to Global Environment. 
Default is "surface". Ignored if \code{as_one_df} is \code{TRUE}.}
}
\description{
Create default AERSCREEN input data frames: 
(1) Control options
(2) Source parameters
(3) Building parameters
(4) Surface characteristics
}
\examples{
input_tbl <- new_aerscreen(as_one_df = TRUE)

new_aerscreen(as_one_df = FALSE, add_to_envir = TRUE, surface = "surface_chars")

input_list <- new_aerscreen(as_one_df = FALSE)
}
\keyword{aerscreen}
\keyword{input}
\keyword{new}
\keyword{tables}

