% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aerscreen_out.R
\name{read_aerscreen_out}
\alias{read_aerscreen_out}
\title{Read AERSCREEN output file}
\usage{
read_aerscreen_out(file = "aerscreen.out")
}
\arguments{
\item{file}{A path to a file or literal data as a character string. Default is "aerscreen.out".}
}
\description{
Read an aerscreen.out file or text into an AERSCREEN results table.
}
\examples{
#read_aerscreen_out(file = "aerscreen.out")
}
\keyword{aerscreen}
\keyword{load}
\keyword{output}
\keyword{read}
\keyword{results}

