% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aerscreen.R
\name{run_aerscreen}
\alias{run_aerscreen}
\title{Run AERSCREEN.exe with designated input file}
\usage{
run_aerscreen(input = aerscreen_inp, output = "aerscreen_results",
  exe_folder = "aerscreen")
}
\arguments{
\item{input}{AERSCREEN input data.frame. Defaults to aerscreen_inp. If a file path is used (e.g. "inputs/aerscreen.inp"), the model will be directed to use the input file.}

\item{output}{File name for model output. Defaults to "aerscreen_results".}

\item{exe_folder}{aerscreen.exe location. Defaults to "aerscreen/".}
}
\description{
Run AERSCREEN.exe with designated input file.
}
\examples{
run_aerscreen(input = "inputs/aerscreen.inp", output = "output_test1")
}
\keyword{aermod}
\keyword{aerscreen}
\keyword{model}
\keyword{run}

