% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aerscreen.R
\name{run_aerscreen}
\alias{run_aerscreen}
\title{Run AERSCREEN}
\usage{
run_aerscreen(input = "aerscreen.inp", out_file = as.character(NA),
  exe_folder = "aerscreen")
}
\arguments{
\item{input}{An AERSCREEN input file or file path. Defaults to "aerscreen.inp". 
File is written to \code{exe_folder} as "aerscreen.inp".}

\item{exe_folder}{Folder path containing \code{aerscreen.exe}.}

\item{results_file}{File name for saving model output. 
If blank, the designated OUT file in the input file will be used.}
}
\description{
Run aerscreen.exe on input file.
}
\examples{
\dontrun{
run_aerscreen(input      = "inputs/aerscreen.inp", 
              out_file   = "output_test1",
              exe_folder = "EPA")
}
}
\keyword{aerscreen}
\keyword{dispersion}
\keyword{model}
\keyword{run}

