% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aerscreen.R
\name{run_aerscreen}
\alias{run_aerscreen}
\title{Run AERSCREEN}
\usage{
run_aerscreen(input = "aerscreen.inp", results_file = as.character(NA),
  exe_folder = "aerscreen/")
}
\arguments{
\item{input}{AERSCREEN input file. Defaults to aerscreen_inp. A file path may be used (ex. "aerscreen.inp").}

\item{results_file}{File name for saving model output. 
If blank, the designated OUT file in the input file will be used.}

\item{exe_folder}{Folder containing \code{aerscreen.exe}.}
}
\description{
Run aerscreen.exe with designated input file.
}
\examples{
#run_aerscreen(input    = "inputs/aerscreen.inp", 
 #              out_file = "output_test1")
}
\keyword{aerscreen}
\keyword{dispersion}
\keyword{model}
\keyword{run}

