% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aerscreen.R
\name{write_aerscreen}
\alias{write_aerscreen}
\title{Save a data frame to an input text file for the AERSCREEN model.}
\usage{
write_aerscreen(x, path)
}
\arguments{
\item{x}{A data frame to write to disk}

\item{path}{Path to write to.}
}
\description{
Save a data frame to an input text file for the AERSCREEN model.
}
\examples{
write_aerscreen(x = aerscreen_inp, path = "Inputs/aerscreen.inp")
}
\keyword{aermod}
\keyword{aerscreen}
\keyword{save}
\keyword{write}

